com.unity.quicksearch@3.0.0-preview.24 ## [3.0.0-preview.24] - 2022-04-20 - Add error reporting when a search index fails to load. - Fix auto-complete text overlapping. - Fix how the global search window is reused in different context. - Fix nested search expression constants evaluation (i.e. constant{...}) - Fix property database concurrency under certain conditions. - Fix query string expression parsing. - Fix scene search proposition enumeration for large scenes. - Fix search indexer concurrent issue when asset are deleted and the index is being updated at the same time. - Fix search picker persisted search flags. - Improve asset provider preview generation. - Update default Temp* pattern exclusion of folders.

+++ Editor/SearchService.cs
@@ -1,9 +1,14 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using UnityEditor.SearchService;
 using UnityEditorInternal;
 using UnityEngine;
 using UnityEngine.Search;
+
+#if USE_SEARCH_ENGINE_API
+using System.Reflection;
+#endif
 
 namespace UnityEditor.Search
 {
@@ -49,9 +54,18 @@ namespace UnityEditor.Search
             }
         }
 
+        #if USE_SEARCH_ENGINE_API
+        internal static List<AdvancedObjectSelector> ObjectSelectors { get; private set; }
+
+        internal static IEnumerable<AdvancedObjectSelector> OrderedObjectSelectors => ObjectSelectors.OrderBy(p => p.priority);
+        #endif
+
         static SearchService()
         {
             Refresh();
+            #if USE_SEARCH_ENGINE_API
+            RefreshObjectSelectors();
+            #endif
             SetupSearchFirstUse();
         }
 
@@ -154,11 +168,70 @@ namespace UnityEditor.Search
             }
         }
 
+        #if USE_SEARCH_ENGINE_API
+        internal static void RefreshObjectSelectors()
+        {
+            var validators = ReflectionUtils.LoadAllMethodsWithAttribute<AdvancedObjectSelectorValidatorAttribute, AdvancedObjectSelectorValidator>(
+                (loaded, mi, attribute, handler) =>
+                    LoadAdvancedObjectSelectorAttribute<AdvancedObjectSelectorValidatorAttribute, AdvancedObjectSelectorValidator, AdvancedObjectSelectorValidatorHandler>(
+                    loaded, mi, attribute, handler, "Advanced Object Selector Validator", (a, h) => GenerateAdvancedObjectSelectorValidatorWrapper(a, h)),
+                MethodSignature.FromDelegate<AdvancedObjectSelectorValidatorHandler>(), ReflectionUtils.AttributeLoaderBehavior.DoNotThrowOnValidation).ToDictionary(validator => validator.id.GetHashCode());
+
+            ObjectSelectors = ReflectionUtils.LoadAllMethodsWithAttribute<AdvancedObjectSelectorAttribute, AdvancedObjectSelector>(
+                (loaded, mi, attribute, handler) =>
+                    LoadAdvancedObjectSelectorAttribute<AdvancedObjectSelectorAttribute, AdvancedObjectSelector, AdvancedObjectSelectorHandler>(
+                    loaded, mi, attribute, handler, "Advanced Object Selector", (a, h) => GenerateAdvancedObjectSelectorWrapper(validators, a, h)),
+                MethodSignature.FromDelegate<AdvancedObjectSelectorHandler>(), ReflectionUtils.AttributeLoaderBehavior.DoNotThrowOnValidation).ToList();
+        }
+
+        static THandlerWrapper LoadAdvancedObjectSelectorAttribute<TAttribute, THandlerWrapper, TDelegate>(IReadOnlyCollection<THandlerWrapper> loaded, MethodInfo mi, TAttribute attribute, Delegate handler, string attributeName, Func<TAttribute, TDelegate, THandlerWrapper> wrapperGenerator)
+        where TAttribute : Attribute, IAdvancedObjectSelectorAttribute
+        where THandlerWrapper : IAdvancedObjectSelector
+        {
+            if (string.IsNullOrEmpty(attribute.id))
+                throw new CustomAttributeFormatException($"Null or empty {attributeName} id for handler \"{ReflectionUtils.GetMethodFullName(mi)}\"");
+
+            if (handler is TDelegate selectorHandler)
+            {
+                if (loaded.Any(p => p.id.Equals(attribute.id, StringComparison.Ordinal)))
+                    throw new CustomAttributeFormatException($"{attributeName} id \"{attribute.id}\" for \"{ReflectionUtils.GetMethodFullName(mi)}\" is already used by another handler.");
+                return wrapperGenerator(attribute, selectorHandler);
+            }
+            throw new CustomAttributeFormatException($"Invalid {attributeName} handler \"{attribute.id}\" using \"{ReflectionUtils.GetMethodFullName(mi)}\"");
+        }
+
+        static AdvancedObjectSelectorValidator GenerateAdvancedObjectSelectorValidatorWrapper(AdvancedObjectSelectorValidatorAttribute attribute, AdvancedObjectSelectorValidatorHandler handler)
+        {
+            return new AdvancedObjectSelectorValidator(attribute.id, handler);
+        }
+
+        static AdvancedObjectSelector GenerateAdvancedObjectSelectorWrapper(Dictionary<int, AdvancedObjectSelectorValidator> validators, AdvancedObjectSelectorAttribute attribute, AdvancedObjectSelectorHandler handler)
+        {
+            if (!validators.TryGetValue(attribute.id.GetHashCode(), out var validator))
+                throw new CustomAttributeFormatException($"Advanced Object Selector id \"{attribute.id}\" does not have a matching validator.");
+
+            var priority = attribute.defaultPriority;
+            var active = attribute.defaultActive;
+            var displayName = string.IsNullOrEmpty(attribute.displayName) ? SearchUtils.ToPascalWithSpaces(attribute.id) : attribute.displayName;
+            if (SearchSettings.TryGetObjectSelectorSettings(attribute.id, out var settings))
+            {
+                priority = settings.priority;
+                active = settings.active;
+            }
+            return new AdvancedObjectSelector(attribute.id, displayName, priority, active, handler, validator);
+        }
+
+        internal static AdvancedObjectSelector GetObjectSelector(string selectorId)
+        {
+            return ObjectSelectors.Find(p => p.id.Equals(selectorId, StringComparison.Ordinal));
+        }
+        #endif
+
         /// <summary>
         /// Create context from a list of provider id.
         /// </summary>
         /// <param name="providerIds">List of provider id</param>
-        /// <param name="searchText">seach Query</param>
+        /// <param name="searchText">Search Query</param>
         /// <param name="flags">Options defining how the query will be performed</param>
         /// <returns>New SearchContext</returns>
         public static SearchContext CreateContext(IEnumerable<string> providerIds, string searchText = "", SearchFlags flags = SearchFlags.Default)
@@ -170,7 +243,7 @@ namespace UnityEditor.Search
         /// Create context from a list of providers.
         /// </summary>
         /// <param name="providers">List of providers</param>
-        /// <param name="searchText">seach Query</param>
+        /// <param name="searchText">Search Query</param>
         /// <param name="flags">Options defining how the query will be performed</param>
         /// <returns>New SearchContext</returns>
         public static SearchContext CreateContext(IEnumerable<SearchProvider> providers, string searchText = "", SearchFlags flags = SearchFlags.Default)
@@ -449,7 +522,9 @@ namespace UnityEditor.Search
             context.asyncItemReceived += ReceiveItems;
             context.sessionStarted += OnSessionStarted;
             context.sessionEnded += OnSessionEnded;
-            GetItems(context, options | SearchFlags.FirstBatchAsync);
+            var firstResults = GetItems(context, options);
+            if (firstResults.Count > 0)
+                ReceiveItems(context, firstResults);
             firstBatchResolved = true;
             if (sessionCount == 0 && !completed)
             {
@@ -757,10 +832,27 @@ namespace UnityEditor.Search
             {
                 onIndexReady?.Invoke(indexName, indexPath.Replace("\\", "/"), () =>
                 {
-                    if (EditorUtility.IsPersistent(db))
-                        Resources.UnloadAsset(db);
-                    if (options.HasNone(IndexingOptions.Keep))
-                        AssetDatabase.DeleteAsset(indexPath);
+                    SearchDatabase.Unload(db);
+
+                    try
+                    {
+                        if (options.HasNone(IndexingOptions.Keep) && System.IO.File.Exists(indexPath))
+                        {
+                            #if USE_SEARCH_MODULE
+                            if (options.HasAny(IndexingOptions.Temporary))
+                                System.IO.File.Delete(indexPath);
+                            else
+                            #endif
+                            { 
+                                AssetDatabase.DeleteAsset(indexPath);
+                            }
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        // Ignore any file IO errors (for the user)
+                        Console.WriteLine(ex.Message);
+                    }
                 });
             }
             else
@@ -810,9 +902,9 @@ namespace UnityEditor.Search
 
         static IEnumerable<SearchItem> EvaluateExpression(SearchExpression expression, SearchContext context)
         {
-            #if USE_PROPERTY_DATABASE
+#if USE_PROPERTY_DATABASE
             using (SearchMonitor.GetView())
-            #endif
+#endif
             {
                 var evaluationFlags = SearchExpressionExecutionFlags.ThreadedEvaluation;
                 var it = expression.Execute(context, evaluationFlags).GetEnumerator();
